Move SSH key generation to separate file
